+++ src/api/auth/protectedRoutesRest.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const protectedRoutesRest = express.Router()
-const httpResponseUtils = require('../../utils/httpResponseUtils')
-const i18nUtils = require('../../utils/i18nUtils')
+const httpResponseUtils = require('../../utils/httpResponse')
+const i18nUtils = require('../../utils/i18n')
 const message = i18nUtils.translate('Hello')
 
 protectedRoutesRest.get('/logger', (req, res) => {
