+++ src/utils/jwt.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken')
-const config = require('../config.json')
+const config = require('../config')
 
 const validateToken = (req) => {
   return new Promise((resolve, reject) => {
